+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # CHANGELOG
 ## master
+- Allow `console.{log, info, warn, error}` override
+
 ## v0.1.0 (2015-09-14)
 - **Initial release**